Move to client side fetching of records list (#45)
Server side rendering for the list of records in a collection makes the
site a target for bots and scrapers with an infinite capacity to DOS the
project. Another neat benefit of CSR in this case is simplified data
fetching logic - no need for an `initialData` prop for building on top
of server fetched data.